+++ src/index.js
@@ -1,5 +1,3 @@
 export * from './di_tokens';
-export * from './local_storage';
-export * from './session_storage';
+export * from './storage';
 export * from './storage_module';
-export * from './web_storage';