+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 This file contains highlights of what changes on each version of the [Web Storage for Angular](https://github.com/cedx/ngx-storage) project.
 
 ## Version 0.2.0
-- Breaking change: renamed the `DOMStorage` class to `WebStorage`.
 - Breaking change: removed the `LocalStorage` and `SessionStorage` classes.
 - Breaking change: removed the `WINDOW` injection token.
+- Breaking change: renamed the `DOMStorage` class to `Storage`.
 - Added the `LocalStorage` and `SessionStorage` injection tokens.
 - Updated the dependencies.